refactor: extract code from state_reconciler
